Move unscaled policy to policy_head
