+++ rl_blox/blox/function_approximator/policy_head.py
@@ -6,7 +6,7 @@ from flax import nnx
 
 
 class DeterministicTanhPolicy(nnx.Module):
-    r"""Deterministic policy represented with a function approximator.
+    r"""Policy with tanh scaling for continuous action spaces.
 
     The deterministic policy directly maps observations to actions, hence,
     represents the function :math:`\pi(o) = a`.
@@ -63,13 +63,18 @@ class StochasticPolicyBase(nnx.Module):
 
 
 class GaussianTanhPolicy(StochasticPolicyBase):
-    r"""Gaussian policy represented with a function approximator.
+    r"""Gaussian policy with tanh scaling for continuous action spaces.
 
     The gaussian policy maps observations to mean and log variance of an
     action, hence, represents the distribution :math:`\pi(a|o)`.
 
     The output of the underlying function approximator is scaled with tanh
     to [-1, 1] and then scaled to [action_space.low, action_space.high].
+
+    Parameters
+    ----------
+    net : nnx.Module
+        Gaussian neural network that maps observations to mean and log variance.
     """
 
     net: nnx.Module
@@ -120,3 +125,92 @@ class GaussianTanhPolicy(StochasticPolicyBase):
         return distrax.MultivariateNormalDiag(
             loc=mean, scale_diag=std
         ).log_prob(action)
+
+
+class GaussianPolicy(StochasticPolicyBase):
+    """Gaussian policy for continuous action spaces.
+
+    Wraps a Gaussian neural network that maps observations to a Gaussian
+    distribution over actions, i.e., mean vector and log variance vector.
+
+    Parameters
+    ----------
+    net : nnx.Module
+        Gaussian neural network that maps observations to mean and log variance.
+    """
+
+    net: nnx.Module
+
+    def __init__(self, net: nnx.Module):
+        self.net = net
+
+    def __call__(self, observation: jnp.ndarray) -> jnp.ndarray:
+        return self.net(observation)[0]
+
+    def sample(self, observation: jnp.ndarray, key: jnp.ndarray) -> jnp.ndarray:
+        """Sample action from Gaussian distribution."""
+        mean, log_var = self.net(observation)
+        log_std = jnp.clip(0.5 * log_var, -20.0, 2.0)
+        std = jnp.exp(log_std)
+        # same as
+        # jax.random.normal(key, mean.shape)
+        # * jnp.exp(jnp.clip(0.5 * log_var, -20.0, 2.0))
+        # + mean
+        return distrax.MultivariateNormalDiag(loc=mean, scale_diag=std).sample(
+            seed=key,
+            sample_shape=(),
+        )
+
+    def log_probability(
+        self,
+        observation: jnp.ndarray,
+        action: jnp.ndarray,
+    ) -> jnp.ndarray:
+        """Compute log probability of action given observation."""
+        mean, log_var = self.net(observation)
+        log_std = jnp.clip(0.5 * log_var, -20.0, 2.0)
+        std = jnp.exp(log_std)
+        # same as
+        # -jnp.log(std)
+        # - 0.5 * jnp.log(2.0 * jnp.pi)
+        # - 0.5 * ((action - mean) / std) ** 2
+        return distrax.MultivariateNormalDiag(
+            loc=mean, scale_diag=std
+        ).log_prob(action)
+
+
+class SoftmaxPolicy(StochasticPolicyBase):
+    r"""Softmax policy for discrete action spaces.
+
+    Wraps a softmax neural network that maps observations to the logits of each
+    action.
+
+    Parameters
+    ----------
+    net : nnx.Module
+        Neural network that maps observations to logits.
+    """
+
+    net: nnx.Module
+
+    def __init__(self, net: nnx.Module):
+        self.net = net
+
+    def __call__(self, observation: jnp.ndarray) -> jnp.ndarray:
+        return nnx.softmax(self.logits(observation))
+
+    def logits(self, observation: jnp.ndarray) -> jnp.ndarray:
+        return self.net(observation)
+
+    def sample(self, observation: jnp.ndarray, key: jnp.ndarray) -> jnp.ndarray:
+        return distrax.Categorical(logits=self.logits(observation)).sample(
+            seed=key,
+            sample_shape=(),
+        )
+
+    def log_probability(
+        self, observation: jnp.ndarray, action: jnp.ndarray
+    ) -> jnp.ndarray:
+        return distrax.Categorical(logits=self.logits(observation)).log_prob(
+            action
+        )
